the gain might not be the issue trying to change to int16

+++ STC_Rover/GUI_Windows/mic_stream_test/mic_stream_test.py
@@ -20,7 +20,7 @@ async def send_audio():
         samplerate=AUDIO_RATE,
         channels=AUDIO_CHANNELS,
         blocksize=BLOCKSIZE,
-        dtype='float32',
+        dtype='int16',
         callback=audio_callback
     )
     stream.start()
